evi/chat-history/evi-python-chat-history/pyproject.toml
[tool.poetry]
name = "evi-python-chat-history-example"
version = "0.1.0"
description = ""
authors = ["zach <hello@hume.ai>"]
readme = "README.md"
# package-mode = "false"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
hume = "^0.7.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "evi-python-chat-history-example.main:main"
